+++ app/src/main/java/com/example/shopping/model/bean/SortItemListBean.java
@@ -124,7 +124,7 @@ public class SortItemListBean {
             private int id;
             private String name;
             private String list_pic_url;
-            private int retail_price;
+            private float retail_price;
 
             public int getId() {
                 return id;
@@ -150,11 +150,11 @@ public class SortItemListBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public int getRetail_price() {
+            public float getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(int retail_price) {
+            public void setRetail_price(float retail_price) {
                 this.retail_price = retail_price;
             }
         }
@@ -206,7 +206,7 @@ public class SortItemListBean {
             private int id;
             private String name;
             private String list_pic_url;
-            private int retail_price;
+            private float retail_price;
 
             public int getId() {
                 return id;
@@ -232,11 +232,11 @@ public class SortItemListBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public int getRetail_price() {
+            public float getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(int retail_price) {
+            public void setRetail_price(float retail_price) {
                 this.retail_price = retail_price;
             }
         }
